requirements.txt
requests
python-dotenv
mkdocs
mkdocs-material